[enh] Migrate label services to new services organization
Bug: https://github.com/eclipse-syson/syson/issues/1628
