+++ backend/services/syson-diagram-services/pom.xml
@@ -62,6 +62,11 @@
 			<artifactId>syson-sysml-metamodel</artifactId>
 			<version>2025.10.1</version>
 		</dependency>
+		<dependency>
+			<groupId>org.eclipse.syson</groupId>
+			<artifactId>syson-services</artifactId>
+			<version>2025.10.1</version>
+		</dependency>
 		<!-- Tests -->
 		<dependency>
 			<groupId>org.testcontainers</groupId>
