+++ backend/services/syson-diagram-services/src/main/java/org/eclipse/syson/diagram/services/aql/DiagramMutationAQLService.java
@@ -12,12 +12,57 @@
  *******************************************************************************/
 package org.eclipse.syson.diagram.services.aql;
 
+import java.util.Objects;
+
+import org.eclipse.syson.diagram.services.DiagramMutationLabelService;
+import org.eclipse.syson.sysml.Element;
+
 /**
  * Entry point for all diagram-related services doing mutations in diagrams and called by AQL expressions in diagram
- * descriptions.
+ * descriptions. This class is not a @Service class but act as it was, because it is called by IJavaServiceProvider.
  *
  * @author arichard
  */
 public class DiagramMutationAQLService {
 
+    private final DiagramMutationLabelService diagramMutationLabelService;
+
+    public DiagramMutationAQLService(DiagramMutationLabelService diagramMutationLabelService) {
+        this.diagramMutationLabelService = Objects.requireNonNull(diagramMutationLabelService);
+    }
+
+    /**
+     * {@link DiagramMutationLabelService#directEdit(Element, String)}.
+     */
+    public Element directEdit(Element element, String newLabel) {
+        return this.diagramMutationLabelService.directEdit(element, newLabel);
+    }
+
+    /**
+     * {@link DiagramMutationLabelService#directEditNode(Element, String)}.
+     */
+    public Element directEditNode(Element element, String newLabel) {
+        return this.diagramMutationLabelService.directEditNode(element, newLabel);
+    }
+
+    /**
+     * {@link DiagramMutationLabelService#directEditListItem(Element, String)}.
+     */
+    public Element directEditListItem(Element element, String newLabel) {
+        return this.diagramMutationLabelService.directEditListItem(element, newLabel);
+    }
+
+    /**
+     * {@link DiagramMutationLabelService#editMultiplicityRangeCenterLabel(Element, String)}.
+     */
+    public Element editMultiplicityRangeCenterLabel(Element element, String newLabel) {
+        return this.diagramMutationLabelService.editMultiplicityRangeCenterLabel(element, newLabel);
+    }
+
+    /**
+     * {@link DiagramMutationLabelService#editEdgeCenterLabel(Element, String)}.
+     */
+    public Element editEdgeCenterLabel(Element element, String newLabel) {
+        return this.diagramMutationLabelService.editEdgeCenterLabel(element, newLabel);
+    }
 }
