+++ backend/services/syson-diagram-services/src/test/java/org/eclipse/syson/diagram/services/DiagramQueryLabelServiceTest.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2024, 2025 Obeo.
+ * Copyright (c) 2025 Obeo.
  * This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v2.0
  * which accompanies this distribution, and is available at
@@ -10,14 +10,13 @@
  * Contributors:
  *     Obeo - initial API and implementation
  *******************************************************************************/
-package org.eclipse.syson.diagram.common.view.services;
+package org.eclipse.syson.diagram.services;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import java.util.List;
 
-import org.eclipse.sirius.components.core.api.IFeedbackMessageService;
 import org.eclipse.syson.sysml.AttributeUsage;
 import org.eclipse.syson.sysml.ConstraintUsage;
 import org.eclipse.syson.sysml.Dependency;
@@ -44,11 +43,11 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 /**
- * View Label services tests.
+ * Label services tests.
  *
  * @author jmallet
  */
-public class ViewLabelServiceTest {
+public class DiagramQueryLabelServiceTest {
 
     private static final String ATTRIBUTE_USAGE_NAME = "myAttributeUsage";
 
@@ -60,11 +59,115 @@ public class ViewLabelServiceTest {
 
     private static final String SHORT_NAME_LABEL = LabelConstants.LESSER_THAN + SHORT_NAME + LabelConstants.GREATER_THAN;
 
-    private ViewLabelService viewLabelService;
+    private DiagramQueryLabelService labelService;
 
     @BeforeEach
     void beforeEach() {
-        this.viewLabelService = new ViewLabelService(new IFeedbackMessageService.NoOp(), new ShowDiagramsIconsService());
+        this.labelService = new DiagramQueryLabelService();
+    }
+
+    @DisplayName("Check Attribute label with name and short name")
+    @Test
+    void testAttributeLabelWithNameAndShortName() {
+        AttributeUsage attributeUsage = SysmlFactory.eINSTANCE.createAttributeUsage();
+        attributeUsage.setDeclaredName(ATTRIBUTE_USAGE_NAME);
+        attributeUsage.setDeclaredShortName(ATTRIBUTE_USAGE_SHORT_NAME);
+        StringBuilder expectedLabel = new StringBuilder();
+        expectedLabel.append(LabelConstants.OPEN_QUOTE)
+                .append("attribute")
+                .append(LabelConstants.CLOSE_QUOTE)
+                .append(LabelConstants.CR)
+                .append(LabelConstants.LESSER_THAN)
+                .append(ATTRIBUTE_USAGE_SHORT_NAME)
+                .append(LabelConstants.GREATER_THAN)
+                .append(LabelConstants.SPACE)
+                .append(ATTRIBUTE_USAGE_NAME);
+        assertEquals(expectedLabel.toString(), this.labelService.getContainerLabel(attributeUsage));
+    }
+
+    @DisplayName("Check Usage prefix label with default properties")
+    @Test
+    void testUsagePrefixLabelWithDefaultProperties() {
+        Usage usageWithNoDirection = SysmlFactory.eINSTANCE.createUsage();
+        assertEquals(LabelConstants.REF + LabelConstants.SPACE, this.labelService.getUsageListItemPrefix(usageWithNoDirection));
+    }
+
+    @DisplayName("Check Usage prefix label with abstract property")
+    @Test
+    void testUsagePrefixLabelWithAbstractProperty() {
+        Usage usageWithNoDirection = SysmlFactory.eINSTANCE.createUsage();
+        usageWithNoDirection.setIsAbstract(true);
+        assertEquals(LabelConstants.ABSTRACT + LabelConstants.SPACE + LabelConstants.REF + LabelConstants.SPACE, this.labelService.getUsageListItemPrefix(usageWithNoDirection));
+    }
+
+    @DisplayName("Check Usage prefix label with derived property")
+    @Test
+    void testUsagePrefixLabelWithDerivedProperty() {
+        Usage usageWithNoDirection = SysmlFactory.eINSTANCE.createUsage();
+        usageWithNoDirection.setIsDerived(true);
+        assertEquals(LabelConstants.DERIVED + LabelConstants.SPACE + LabelConstants.REF + LabelConstants.SPACE, this.labelService.getUsageListItemPrefix(usageWithNoDirection));
+    }
+
+    @DisplayName("Check Usage prefix label with end property")
+    @Test
+    void testUsagePrefixLabelWithEndProperty() {
+        Usage usageWithNoDirection = SysmlFactory.eINSTANCE.createUsage();
+        usageWithNoDirection.setIsEnd(true);
+        assertEquals(LabelConstants.END + LabelConstants.SPACE + LabelConstants.REF + LabelConstants.SPACE, this.labelService.getUsageListItemPrefix(usageWithNoDirection));
+    }
+
+    @DisplayName("Check Usage prefix label with constant property")
+    @Test
+    void testUsagePrefixLabelWithConstantProperty() {
+        Usage usageWithNoDirection = SysmlFactory.eINSTANCE.createUsage();
+        usageWithNoDirection.setIsConstant(true);
+        assertEquals(LabelConstants.CONSTANT + LabelConstants.SPACE + LabelConstants.REF + LabelConstants.SPACE, this.labelService.getUsageListItemPrefix(usageWithNoDirection));
+    }
+
+    @DisplayName("Check Usage prefix label with readonly property")
+    @Test
+    void testUsagePrefixLabelWithVariationProperty() {
+        Usage usageWithNoDirection = SysmlFactory.eINSTANCE.createUsage();
+        usageWithNoDirection.setIsVariation(true);
+        assertEquals(LabelConstants.VARIATION + LabelConstants.SPACE + LabelConstants.REF + LabelConstants.SPACE, this.labelService.getUsageListItemPrefix(usageWithNoDirection));
+    }
+
+    @DisplayName("Check Usage prefix label with many custom properties")
+    @Test
+    void testUsagePrefixLabelWithManyCustomProperties() {
+        Usage usageWithNoDirection = SysmlFactory.eINSTANCE.createUsage();
+        usageWithNoDirection.setIsAbstract(true);
+        usageWithNoDirection.setIsDerived(true);
+        usageWithNoDirection.setIsEnd(true);
+        usageWithNoDirection.setIsConstant(true);
+        usageWithNoDirection.setIsVariation(true);
+
+        assertEquals(LabelConstants.VARIATION + LabelConstants.SPACE + LabelConstants.CONSTANT + LabelConstants.SPACE
+                + LabelConstants.DERIVED + LabelConstants.SPACE
+                + LabelConstants.END
+                + LabelConstants.SPACE + LabelConstants.REF + LabelConstants.SPACE, this.labelService.getUsageListItemPrefix(usageWithNoDirection));
+    }
+
+    @DisplayName("Check Usage prefix label with direction properties")
+    @Test
+    void testUsagePrefixLabelWithDirectionProperties() {
+        Usage usageWithDirection = SysmlFactory.eINSTANCE.createUsage();
+
+        usageWithDirection.setDirection(FeatureDirectionKind.IN);
+        assertEquals(LabelConstants.IN + LabelConstants.SPACE + LabelConstants.REF + LabelConstants.SPACE, this.labelService.getUsageListItemPrefix(usageWithDirection));
+
+        usageWithDirection.setDirection(FeatureDirectionKind.OUT);
+        assertEquals(LabelConstants.OUT + LabelConstants.SPACE + LabelConstants.REF + LabelConstants.SPACE, this.labelService.getUsageListItemPrefix(usageWithDirection));
+
+        usageWithDirection.setDirection(FeatureDirectionKind.INOUT);
+        assertEquals(LabelConstants.INOUT + LabelConstants.SPACE + LabelConstants.REF + LabelConstants.SPACE, this.labelService.getUsageListItemPrefix(usageWithDirection));
+    }
+
+    @DisplayName("Check AttributeUsage prefix label with default properties")
+    @Test
+    void testAttributeUsagePrefixLabelWithDefaultProperties() {
+        AttributeUsage attributeUsageWithNoDirection = SysmlFactory.eINSTANCE.createAttributeUsage();
+        assertEquals("", this.labelService.getUsageListItemPrefix(attributeUsageWithNoDirection));
     }
 
     @DisplayName("Check Attribute Usage item label with name")
@@ -72,7 +175,7 @@ public class ViewLabelServiceTest {
     void testItemCompartmentLabelWithName() {
         AttributeUsage attributeUsage = SysmlFactory.eINSTANCE.createAttributeUsage();
         attributeUsage.setDeclaredName(ATTRIBUTE_USAGE_NAME);
-        assertEquals(ATTRIBUTE_USAGE_NAME, this.viewLabelService.getCompartmentItemLabel(attributeUsage));
+        assertEquals(ATTRIBUTE_USAGE_NAME, this.labelService.getCompartmentItemLabel(attributeUsage));
     }
 
     @DisplayName("Check Attribute Usage item label with name and short name")
@@ -81,7 +184,7 @@ public class ViewLabelServiceTest {
         attributeUsage.setDeclaredName(ATTRIBUTE_USAGE_NAME);
         attributeUsage.setDeclaredShortName(ATTRIBUTE_USAGE_SHORT_NAME);
         assertEquals(LabelConstants.LESSER_THAN + ATTRIBUTE_USAGE_SHORT_NAME + LabelConstants.GREATER_THAN + LabelConstants.SPACE + ATTRIBUTE_USAGE_NAME,
-                this.viewLabelService.getCompartmentItemLabel(attributeUsage));
+                this.labelService.getCompartmentItemLabel(attributeUsage));
     }
 
     @DisplayName("GIVEN a FeatureValue to an attribute, WHEN it is an initial FeatureValue relationship, THEN the label should use the symbole ':=' instead of '='")
@@ -96,11 +199,11 @@ public class ViewLabelServiceTest {
         featureValue.getOwnedRelatedElement().add(literal);
         attributeUsage.getOwnedRelationship().add(featureValue);
 
-        assertEquals("myAttributeUsage = 1", this.viewLabelService.getCompartmentItemLabel(attributeUsage));
+        assertEquals("myAttributeUsage = 1", this.labelService.getCompartmentItemLabel(attributeUsage));
 
         featureValue.setIsInitial(true);
 
-        assertEquals("myAttributeUsage := 1", this.viewLabelService.getCompartmentItemLabel(attributeUsage));
+        assertEquals("myAttributeUsage := 1", this.labelService.getCompartmentItemLabel(attributeUsage));
 
     }
 
@@ -116,15 +219,15 @@ public class ViewLabelServiceTest {
         featureValue.getOwnedRelatedElement().add(literal);
         attributeUsage.getOwnedRelationship().add(featureValue);
 
-        assertEquals("myAttributeUsage = 1", this.viewLabelService.getCompartmentItemLabel(attributeUsage));
+        assertEquals("myAttributeUsage = 1", this.labelService.getCompartmentItemLabel(attributeUsage));
 
         featureValue.setIsDefault(true);
 
-        assertEquals("myAttributeUsage default = 1", this.viewLabelService.getCompartmentItemLabel(attributeUsage));
+        assertEquals("myAttributeUsage default = 1", this.labelService.getCompartmentItemLabel(attributeUsage));
 
         featureValue.setIsInitial(true);
 
-        assertEquals("myAttributeUsage default := 1", this.viewLabelService.getCompartmentItemLabel(attributeUsage));
+        assertEquals("myAttributeUsage default := 1", this.labelService.getCompartmentItemLabel(attributeUsage));
 
     }
 
@@ -132,7 +235,7 @@ public class ViewLabelServiceTest {
     @Test
     void testItemCompartmentLabelWithoutName() {
         AttributeUsage attributeUsage = SysmlFactory.eINSTANCE.createAttributeUsage();
-        assertEquals("", this.viewLabelService.getCompartmentItemLabel(attributeUsage));
+        assertEquals("", this.labelService.getCompartmentItemLabel(attributeUsage));
     }
 
     @DisplayName("Check Attribute Usage item label with prefix")
@@ -142,7 +245,7 @@ public class ViewLabelServiceTest {
         attributeUsage.setDeclaredName(ATTRIBUTE_USAGE_NAME);
 
         attributeUsage.setIsAbstract(true);
-        assertEquals(LabelConstants.ABSTRACT + LabelConstants.SPACE + ATTRIBUTE_USAGE_NAME, this.viewLabelService.getCompartmentItemLabel(attributeUsage));
+        assertEquals(LabelConstants.ABSTRACT + LabelConstants.SPACE + ATTRIBUTE_USAGE_NAME, this.labelService.getCompartmentItemLabel(attributeUsage));
     }
 
     @DisplayName("Check Attribute Usage item label with multiplicity")
@@ -152,7 +255,7 @@ public class ViewLabelServiceTest {
         attributeUsage.setDeclaredName(ATTRIBUTE_USAGE_NAME);
 
         attributeUsage.setIsOrdered(true);
-        assertEquals(ATTRIBUTE_USAGE_NAME + LabelConstants.SPACE + LabelConstants.ORDERED, this.viewLabelService.getCompartmentItemLabel(attributeUsage));
+        assertEquals(ATTRIBUTE_USAGE_NAME + LabelConstants.SPACE + LabelConstants.ORDERED, this.labelService.getCompartmentItemLabel(attributeUsage));
     }
 
     @DisplayName("Check Attribute Usage item label with prefix and multiplicity")
@@ -164,7 +267,7 @@ public class ViewLabelServiceTest {
         attributeUsage.setIsOrdered(true);
         attributeUsage.setIsAbstract(true);
         assertEquals(LabelConstants.ABSTRACT + LabelConstants.SPACE + ATTRIBUTE_USAGE_NAME + LabelConstants.SPACE + LabelConstants.ORDERED,
-                this.viewLabelService.getCompartmentItemLabel(attributeUsage));
+                this.labelService.getCompartmentItemLabel(attributeUsage));
     }
 
     @DisplayName("GIVEN a ConstraintUsage with no expression, WHEN its label is computed, THEN the label contains the name of the constraint")
@@ -175,7 +278,7 @@ public class ViewLabelServiceTest {
         // Constraints have a special label when they are inside a RequirementConstraintMembership
         RequirementConstraintMembership requirementConstraintMembership = SysmlFactory.eINSTANCE.createRequirementConstraintMembership();
         requirementConstraintMembership.getOwnedRelatedElement().add(constraintUsage);
-        assertThat(this.viewLabelService.getCompartmentItemLabel(constraintUsage)).isEqualTo(CONSTRAINT_USAGE_NAME);
+        assertThat(this.labelService.getCompartmentItemLabel(constraintUsage)).isEqualTo(CONSTRAINT_USAGE_NAME);
     }
 
     @DisplayName("GIVEN a ConstraintUsage with a boolean expression, WHEN its label is computed, THEN the label represents the expression")
@@ -194,7 +297,7 @@ public class ViewLabelServiceTest {
         literalInteger.setValue(2);
         yValue.getOwnedRelatedElement().add(literalInteger);
 
-        assertThat(this.viewLabelService.getCompartmentItemLabel(constraintUsage)).isEqualTo("1 >= 2");
+        assertThat(this.labelService.getCompartmentItemLabel(constraintUsage)).isEqualTo("1 >= 2");
     }
 
     @DisplayName("GIVEN a ConstraintUsage with an expression containing a subject reference, WHEN its label is computed, THEN the label represents the expression")
@@ -217,7 +320,7 @@ public class ViewLabelServiceTest {
         yFeatureReference.getOwnedRelationship().add(yFeatureReferenceMembership);
         yFeatureReferenceMembership.setMemberElement(subjectReference);
 
-        assertThat(this.viewLabelService.getCompartmentItemLabel(constraintUsage)).isEqualTo("1 >= mySubject");
+        assertThat(this.labelService.getCompartmentItemLabel(constraintUsage)).isEqualTo("1 >= mySubject");
 
     }
 
@@ -242,7 +345,7 @@ public class ViewLabelServiceTest {
         yFeatureReference.getOwnedRelationship().add(yFeatureReferenceMembership);
         yFeatureReferenceMembership.setMemberElement(attributeUsage);
 
-        assertThat(this.viewLabelService.getCompartmentItemLabel(constraintUsage)).isEqualTo("1 >= myAttribute");
+        assertThat(this.labelService.getCompartmentItemLabel(constraintUsage)).isEqualTo("1 >= myAttribute");
     }
 
     @DisplayName("GIVEN a ConstraintUsage with an expression containing a single feature chaining, WHEN its label is computed, THEN the label represents the expression")
@@ -267,7 +370,7 @@ public class ViewLabelServiceTest {
         featureChainExpression.getOwnedRelationship().add(featureChainMembership);
         featureChainMembership.setMemberElement(subAttributeUsage);
 
-        assertThat(this.viewLabelService.getCompartmentItemLabel(constraintUsage)).isEqualTo("1 >= myAttribute.mySubAttribute");
+        assertThat(this.labelService.getCompartmentItemLabel(constraintUsage)).isEqualTo("1 >= myAttribute.mySubAttribute");
     }
 
     @DisplayName("GIVEN a ConstraintUsage with an expression containing multiple feature chainings, WHEN its label is computed, THEN the label represents the expression")
@@ -300,7 +403,7 @@ public class ViewLabelServiceTest {
         featureChaining2.setChainingFeature(zAttributeUsage);
         feature.getOwnedRelationship().addAll(List.of(featureChaining1, featureChaining2));
 
-        assertThat(this.viewLabelService.getCompartmentItemLabel(constraintUsage)).isEqualTo("1 >= x.y.z");
+        assertThat(this.labelService.getCompartmentItemLabel(constraintUsage)).isEqualTo("1 >= x.y.z");
     }
 
     @DisplayName("GIVEN a Dependency with a name and short name, WHEN its edge label is computed, THEN the label contains the name and short name")
@@ -310,7 +413,7 @@ public class ViewLabelServiceTest {
         dependency.setDeclaredShortName(SHORT_NAME);
         dependency.setDeclaredName("dependency");
 
-        assertThat(this.viewLabelService.getDependencyLabel(dependency)).isEqualTo(SHORT_NAME_LABEL + " dependency");
+        assertThat(this.labelService.getDependencyLabel(dependency)).isEqualTo(SHORT_NAME_LABEL + " dependency");
     }
 
     @DisplayName("GIVEN an Interface with a name and short name, WHEN its edge label is computed, THEN the label contains the name and short name")
@@ -320,7 +423,7 @@ public class ViewLabelServiceTest {
         interfaceUsage.setDeclaredShortName(SHORT_NAME);
         interfaceUsage.setDeclaredName("interface");
 
-        assertThat(this.viewLabelService.getEdgeLabel(interfaceUsage)).isEqualTo(SHORT_NAME_LABEL + " interface");
+        assertThat(this.labelService.getEdgeLabel(interfaceUsage)).isEqualTo(SHORT_NAME_LABEL + " interface");
     }
 
     /**
