+++ scripts/check-coverage.jsh
@@ -35,20 +35,20 @@ var moduleCoverageData = List.of(
   new ModuleCoverage("syson-sysml-metamodel-edit", 16.0),
   new ModuleCoverage("syson-siriusweb-customnodes-metamodel", 44.0),
   new ModuleCoverage("syson-siriusweb-customnodes-metamodel-edit", 0.0),
-  new ModuleCoverage("syson-direct-edit-grammar", 66.0),
-  new ModuleCoverage("syson-services", 72.0),
+  new ModuleCoverage("syson-direct-edit-grammar", 67.0),
+  new ModuleCoverage("syson-services", 67.0),
   new ModuleCoverage("syson-sysml-rest-api-services", 93.0),
   new ModuleCoverage("syson-sysml-import", 84.0),
   new ModuleCoverage("syson-sysml-export", 64.0),
   new ModuleCoverage("syson-sysml-validation", 99.0),
   new ModuleCoverage("syson-common-view", 100.0),
-  new ModuleCoverage("syson-diagram-common-view", 85.0),
+  new ModuleCoverage("syson-diagram-common-view", 86.0),
   new ModuleCoverage("syson-standard-diagrams-view", 97.0),
   new ModuleCoverage("syson-table-requirements-view", 77.0),
   new ModuleCoverage("syson-tree-explorer-view", 84.0),
   new ModuleCoverage("syson-application-configuration", 71.0),
   new ModuleCoverage("syson-application", 37.0),
-  new ModuleCoverage("syson-diagram-services", 100.0),
+  new ModuleCoverage("syson-diagram-services", 85.0),
   new ModuleCoverage("syson-form-services", 100.0),
   new ModuleCoverage("syson-model-services", 100.0),
   new ModuleCoverage("syson-representation-services", 100.0),
